Fix product brand cascade, customer VAT, sale decimals, and supplier dropdown.
Load product form JS via the correct base template blocks, add customer VAT number, allow decimal quantities and amounts on new sale, and constrain supplier select width.

+++ transactions/models.py
@@ -169,7 +169,7 @@ class SaleDetail(models.Model):
         max_digits=10,
         decimal_places=2
     )
-    quantity = models.PositiveIntegerField()
+    quantity = models.DecimalField(max_digits=12, decimal_places=3)
     total_detail = models.DecimalField(max_digits=10, decimal_places=2)
 
     class Meta:
